// src/shared/domain/types/expense.ts
export type ExpenseCategory =
  // SHOPPING
  | "FOOD"
  | "TAKEAWAY"
  | "ALCOHOL"
  // APARTMENT
  | "RENT"
  | "UTILITIES"
  | "MAINTENANCE"
  // TRANSPORT
  | "PUBLIC_TRANSPORT"
  | "FUEL"
  | "CAR"
  // HEALTHCARE
  | "DOCTOR"
  | "PHARMACY"
  // ENTERTAINMENT
  | "GYM"
  | "CINEMA"
  | "BOOKS"
  // OTHER
  | "OTHER";